@@ -21,7 +21,7 @@
     "create": "node ./scripts/create-rule",
     "flow": "flow",
     "lint:fix": "npm run lint -- --fix",
-    "lint": "eslint  --config .eslintrc src/**/*.ts __tests__ scripts",
+    "lint": "eslint src/**/*.ts __tests__ scripts",
     "prepublishOnly": "yarn build",
     "pretest": "npm run lint:fix && npm run flow",
     "test": "npm run jest",
@@ -36,10 +36,12 @@
     "@babel/preset-env": "^7.15.8",
     "@changesets/cli": "^2.27.7",
     "@svitejs/changesets-changelog-github-compact": "^1.1.0",
+    "@typescript-eslint/eslint-plugin": "^8.32.0",
+    "@typescript-eslint/parser": "^8.32.0",
     "babel-jest": "^27.3.1",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint": "^9.26.0",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-prettier": "^5.4.0",
     "expect": "^27.3.1",
     "flow-bin": "^0.137.0",
     "in-publish": "^2.0.1",
@@ -47,7 +49,7 @@
     "jscodeshift": "^0.13.0",
     "minimist": "^1.2.5",
     "object.assign": "^4.1.2",
-    "prettier": "^2.4.1",
+    "prettier": "^3.5.3",
     "rimraf": "^3.0.2",
     "safe-publish-latest": "^2.0.0",
     "to-ast": "^1.0.0"
