@@ -19,7 +19,7 @@ const validValues = [
   'radiobutton_unchecked',
 ];
 
-module.exports = createValidPropRule(
+export default createValidPropRule(
   'accessibilityComponentType',
   validValues,
   errorMessage
