@@ -32,7 +32,7 @@ const validValues = [
   'pageTurn',
 ];
 
-module.exports = createValidPropRule(
+export default createValidPropRule(
   'accessibilityTraits',
   validValues,
   errorMessage
